KXI-17865: Remove error when upgrading insights without any assemblies running

+++ kxicli/commands/install.py
@@ -385,7 +385,7 @@ def perform_upgrade(namespace, release, chart_repo_name, assembly_backup_filepat
                                                 chart_repo_name, install_operator, is_op_upgrade, crd_data)
 
     click.secho(phrases.upgrade_asm_reapply, bold=True)
-    if all(assembly._create_assemblies_from_file(namespace=namespace, filepath=assembly_backup_filepath)):
+    if assembly_backup_filepath and all(assembly._create_assemblies_from_file(namespace=namespace, filepath=assembly_backup_filepath)):
         os.remove(assembly_backup_filepath)
 
     if upgraded:
